Fixes SPELL_AURA_MOD_RESISTANCE (and others) with negative level scaling.
As described in issue #30 various aura mods like
SPELL_AURA_MOD_RESISTANCE are calculated based on base points. But
sometimes the level scaling is defined as negative, but the level is
either clamped at 0 or not initialized or something. Resulting in
debuffs being buffs.

This fixes that behavior and many other behaviors.